refactor(serve): keep repolint budgets while adding the Host guard
Move hostGuard, hostGateState, setListenAddr, and csrfGuard into the new
internal/serve/hostguard.go and condense the added comment blocks in
gitcmd, installsource, permission, and preview so every repolint essay and
file-size budget stays at its existing baseline. serve.go ends 17 lines
smaller than main-v2. Behavior is unchanged; the moved guards keep their
tests and wiring.

Verification: go run ./tools/repolint -> clean (1257 baselined findings);
go test ./internal/serve ./internal/gitcmd ./internal/installsource
./internal/permission ./internal/tool/builtin all pass.
